feat: Add CSS Tree as a npm dependency
CSS Tree is a set of tools for working with CSS, including fast detailed
parser (string->AST), walkers, generators (AST->string) and even lexer
(validation and matching) based on knowledge of spec and browser
implementations. The main goal is to be efficient and W3C spec compliant,
with focus on analyzing and source-to-source processing.