refactor: Activate extension when a CSS file is opened
Before, this extension was authored in such a way that it included a 'Say
Hello' command. By invoking this command this extension got activated. For
obvious reasons this is not necessary anymore. Now, this extension is
activated when a CSS file is encountered (by opening a CSS file or by
navigating to a tab that is linked to a CSS file).